+++ docs/whatsnew.rst
@@ -5,6 +5,7 @@ Version 0.3.4
 `````````````
 
 - API: Added a ``warning`` option for controling whether stderr is included in the output.
+- API: Changed the ``on_error`` option to ``error`` for compatability with knitr and symmytry with the ``warning`` option.
 
 Version 0.3.3
 `````````````